fix: add detection to workspace'd kit - in a very ugly way
